refactor(audit): split god-files to clear structural findings (partial, #5645)
